added authorization with games

+++ MiniAPI.Services/Games/Commands/DeleteGame.cs
@@ -13,6 +13,8 @@ public class DeleteGame
     public class Command : IRequest<Guid>
     {
         public required Guid Id { get; set; }
+
+        public required Guid UserId { get; set; }
     }
 
     public class Handler(MiniAPIContext context) : IRequestHandler<Command, Guid>
@@ -21,6 +23,10 @@ public class DeleteGame
         {
             Game? game = await context.Games.FirstOrDefaultAsync(g => g.GameID == request.Id,
              ct) ?? throw new Exception("Game does not exist.");
+            if (game.OwnerID != request.UserId)
+            {
+                throw new Exception("Only the owner can modify game.");
+            }
             context.Games.Remove(game);
             await context.SaveChangesAsync(ct);
             return request.Id;
